new api test class to refactor all tests. Resolves #497
