+++ test/frontend/MembershipPlan.test.js
@@ -9,7 +9,7 @@ import { membershipPlans } from '../../src/Enums';
 import MembershipPlan from
   '../../src/Pages/MembershipApplication/MembershipPlan';
 import Adapter from 'enzyme-adapter-react-16';
-import { mockMonth, revertClock } from '../mocks/Date';
+import { mockMonth, revertClock } from '../util/mocks/Date';
 
 Enzyme.configure({ adapter: new Adapter() });
