+++ test/frontend/UserProfile.test.js
@@ -7,7 +7,7 @@ import 'regenerator-runtime/runtime';
 
 import UserProfile from '../../src/Pages/Profile/MemberView/Profile';
 import InfoCard from '../../src/Pages/Profile/MemberView/InfoCard';
-import { mockMonth, revertClock } from '../mocks/Date';
+import { mockMonth, revertClock } from '../util/mocks/Date';
 import Adapter from 'enzyme-adapter-react-16';
 
 import winter from '../../src/Pages/Profile/MemberView/Image/winter';
