+++ test/officerManager.js
@@ -16,13 +16,13 @@ const {
 let app = null;
 const expect = chai.expect;
 // tools for testing
-const tools = require('../util/testing-utils/tools.js');
+const tools = require('./util/tools/tools.js');
 const {
   setTokenStatus,
   resetMock,
   restoreMock,
   initializeMock
-} = require('./mocks/TokenValidFunctions');
+} = require('./util/mocks/TokenValidFunctions');
 const { DEFAULT_PHOTO_URL } = require('../api/constants');
 chai.should();
 chai.use(chaiHttp);
